.archive/echo-v1/echo/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
litellm==1.15.0
python-dotenv==1.0.0
groq
python-multipart
